requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==3.2.15
django-extra-fields==3.0.2
djangorestframework==3.15.1
drf-extra-fields==3.7.0
filetype==1.2.0
pillow==10.3.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.2.0
sqlparse==0.5.0
tzdata==2024.1
